backend/requirements.txt
Flask
flask-cors
torch
transformers
google-generativeai
Pillow
pdfplumber
